requirements.txt
alembic
bcrypt
click
Flask
Flask-Bcrypt
Flask-Cors
Flask-Migrate
Flask-SocketIO
Flask-SQLAlchemy
greenlet
gunicorn
eventlet
psycopg2-binary
python-dotenv
google-generativeai
SQLAlchemy
Werkzeug